+++ examples/package.json
@@ -8,9 +8,10 @@
   "license": "MIT",
   "dependencies": {
     "@ant-design/icons": "4.2.2",
-    "@moxb/antd": "0.2.0-beta.41",
-    "@moxb/moxb": "0.2.0-beta.41",
-    "@moxb/semui": "0.2.0-beta.41",
+    "@moxb/antd": "0.2.0-beta.50",
+    "@moxb/html": "0.2.0-beta.50",
+    "@moxb/moxb": "0.2.0-beta.50",
+    "@moxb/semui": "0.2.0-beta.50",
     "antd": "4.6.3",
     "less": "3.12.2",
     "marked": "1.1.1",
