+++ packages/antd/src/routing/NavStepsAnt.tsx
@@ -9,7 +9,7 @@ import { Steps } from 'antd';
 import { StepsProps } from 'antd/lib/steps';
 import { observer } from 'mobx-react';
 import * as React from 'react';
-import { UIFragment, UIFragmentSpec } from '../not-antd';
+import { UIFragment, UIFragmentSpec } from '@moxb/html';
 
 export { StepsProps } from 'antd/lib/steps';
