+++ packages/html/src/LocationDependentArea.tsx
