+++ packages/html/src/index.ts
@@ -1,2 +1,13 @@
 export * from './ActionHtml';
 export * from './BindHtml';
+
+export { LocationDependentArea, LocationDependentAreaProps } from './LocationDependentArea';
+export { UIFragment, renderUIFragment } from './UIFragment';
+export { UIFragmentSpec, extractUIFragmentFromSpec } from './UIFragmentSpec';
+export { Anchor } from './Anchor';
+export { rootOrDetails, DetailProps } from './rootOrDetails';
+export { renderFallback } from './rendering';
+export { Redirect, redirect, redirectTo, redirectToNavRef } from './Redirect';
+export { NavRefRedirect } from './NavRefRedirect';
+export { NavLink, NavLinkProps, NavLinkParams } from './NavLink';
+export { BoundNavLink } from './BoundNavLink';
